+++ lib/autentificacion/signup.dart
@@ -286,7 +286,6 @@ class _SignupState extends State<Signup> {
                                                     flex: 6,
                                                     child: ElevatedButton(
                                                         onPressed: () async {
-                                                          // TODO: comprobacion registro
                                                           if ( await _authService.signup(_email.text, _pwd.text, _name.text)){
                                                             Navigator.of(context).pushNamedAndRemoveUntil('/', (Route<dynamic> route) => false);
                                                           }
